Fix the bug 'weather-data-not-showing'
Fixed the bug which resulted in weather-data not showing on a single
click but double click. It was happening because toggle method was used
to apply block class to 'data-container'. So, one click would add the
class but another would remove it. Solved it by using add method instead
of toggle.